+++ python/examples/keras_classifier.py
@@ -1,9 +1,7 @@
-from keras import Sequential
-from keras import layers
-from keras import utils
+from keras import Sequential, layers, utils
+
 from shaprpy import explain
 from shaprpy.datasets import load_binary_iris
-
 
 dfx_train, dfx_test, dfy_train, dfy_test = load_binary_iris()